Add files via upload
I have made a very basic website using html+css only. No javascript , bootstrap is used here. This is only a front end development. This website contains menus, logo of hospital , a simple good loking picture of a doctor and several buttons. Except the image of the doctor all are created manually.  All the files are provided in the repository along with source code and screenshot (Hospital_website.png) of the output website.